2974 handle missing metadata in plugins (#1064)
* Use Json Schema to validate video plugin data

* Refactor validation into hook

* Add fallback component for schema validation
